+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Core: Updated German translation thanks to University of Greifswald
 - DAV: Remove obsolete DavTask model
 - Core: update cache on setting default user groups
+- Calendar: only display and enable 'Add download link' button if end user has 'access_main' permission
 
 17-03-2025: 6.8.110
 - Finance: Try to sanitize HTML to avoid too much vertical space between text on the PDF